.rtorrent.rc: fix moving of completed torrents
It broke suddenly, probably because the upgrade to 0.9.4 resulted in check_hash
not being ran in finished downloads. The previous config was depending on this,
since it was moving the files as soon as check_hash was finished. The new
config that moves completed torrents to a different directory is taken from:
http://wiki.rtorrent.org/Recipe/Move%20downloaded%20data%20to%20another%20directory%20and%20continue%20seeding