.zshrc: Replace keychain with plain gpg-agent
The ssh-agent is gone, the ssh keys are now handled by the gpg-agent.
The command `gpg-connect-agent -q /bye` on the .zshrc makes sure that
gpg agent starts on every session. This way gpg-agent runs always and
serves both SSH and GPG keys.

The sshcontrol key lists the SSH keys, and since the SSH keys are
different on my laptops, I'm removing that file from version control.